Handle the case of an application freeing host_ptr before we copied it on the device.
This fixes a reflection made on steckdenis.wordpress.com :

"One problem I currently think about is that we can pass a host_ptr to
clCreateBuffer. This pointer contains data to be uploaded to the buffer,
to pre-populate it. The problem is that if we create the buffer after
the clCreateBuffer call, for instance in a clEnqueueNDRangeKernel, how
to be sure that the memory once pointed by host_ptr is still valid ?"

Also, the implementation of buffer->size() is a bit cleaner.